Major update to two-layer model interface physics. Results are now physically releastic, benchmarked by comparing very optically thin dust coatings to a non-layered non-RTE model.
Fixed bug in lower boundary temperature of DISORT model, which was previously
mistakenly set to the base temperature of the entire column, rather than
the dust/rock interface temperature.

Added radiance to effective temperature for disort single-wave, but the
results are unusual. Unsure why. Maybe best to keep using emissT for now,
or write a better calculation in the future that uses emissT + the upwards
scattering flux from the top layer.

Confirmed agreement between Hapke and DISORT boundary fluxes.

Updated auto time steppinng criterion, as it was found that the time steps
were much too large for high-thermal inertia scenarios.

Updated scripts for computing the effective TI of dust layers.