Added depth and temperature-dependent properties for lunar cases.
Switched flux-divergence calculation method for
disort after it was found in slow lunar cases that
the previous method led to unreasonable heat propogation
deep in the subsurface.

Fixed minor issue in grid creating (geometric spacing)
where the grid nodes may not have been exatly
in the center of each layer. Changes appear to be
negligible.